requirements.txt
opencv-python>=4.8.0
python-pptx>=0.6.21
pytesseract>=0.3.10
numpy>=1.24.0
Pillow>=10.0.0 